ccs/requirements.txt
jinja2
pillow
django-google-maps
